requirements.txt
# Core dependencies for PINN project
# Note: ANDES is not included here as it's typically installed separately via conda or pip
# See docs/guides/ANDES_VERSION_INFO.md for ANDES installation and version requirements
# - ANDES 1.10.0 requires Python >= 3.10
# - ANDES 1.9.3 works with Python >= 3.8
# See requirements-andes.txt for ANDES installation options

kvxopt>=1.3.2.0
numpy
# Note: scipy version constraint may conflict with ANDES 1.10.0 (requires scipy<=1.16.3)
# If using ANDES 1.10.0, you may need: scipy>=1.13,<1.17
# For ANDES 1.9.3, this constraint is fine
scipy<1.14
sympy>=1.6,!=1.10.0
pandas
matplotlib
seaborn>=0.13.0
scikit-learn
openpyxl
xlsxwriter
dill
pathos
tqdm
pyyaml
coloredlogs
chardet
psutil
texttable
numba
